requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
altgraph==0.17.4
async-timeout==4.0.3
attrs==23.2.0
black==24.1.0
click==8.1.7
discord.py==2.3.2
frozenlist==1.4.1
idna==3.6
importlib-metadata==7.0.1
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2024.0
python-dotenv==1.0.1
tomli==2.0.1
typing_extensions==4.9.0
yarl==1.9.4
zipp==3.17.0
